feat: add TimeSeriesChart component and integrate it into AccountInfoPanel
